backend-101/lesson-5/pyproject.toml
[project]
name = "lesson-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Anatoliy",email = "bullet2018@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
